Version 1.3: Final version before GUI creation
Removing a different set of K nodes is again the method to look for
a puzzle with only one solution.  After 50 checks a new puzzle is
generated and this process can occur up to 6 times (AKA 300 checks).
The elapsed time of the program was added and some code cleanup as well.